changed model files to be in their own files instead of one
