+++ app/models/__init__.py
@@ -1,3 +1,8 @@
 from .db import db
 from .user import User
+from .payment import Payment
+from .review import Review
+from .task import Task
+from .taskertasktype import TaskerTaskType
+from .tasktype import TaskType
 from .db import environment, SCHEMA
