+++ app/models/tasktype.py
@@ -1,0 +1,18 @@
+from .db import db, environment, SCHEMA, add_prefix_for_prod
+from werkzeug.security import generate_password_hash, check_password_hash
+from flask_login import UserMixin
+from datetime import datetime
+
+class TaskType(db.Model):
+    __tablename__ = 'TaskType'
+
+    if environment == "production":
+        __table_args__ = {'schema': SCHEMA}
+
+    id = db.Column(db.Integer, primary_key=True)
+    type = db.Column(db.Text)
+    createdAt = db.Column(db.DateTime, default=datetime.utcnow)
+    updatedAt = db.Column(db.DateTime, default=datetime.utcnow)
+
+    tasks = db.relationship('Task', back_populates='taskType')
+    taskerTaskTypes = db.relationship('TaskerTaskType', back_populates='taskType')
